// client_angular/src/app/Model/model.ts
export class User {
  username: string;
  email: string;
  password: string;
  password_conf: string;
}
export class Contact {
  name: string;
  address: string;
  email: string;
  phone_number: string;
  subject: string;
  message: string;
}
export class Portfolio {
  photo: string;
  content: string;
  in_front: string;
}